refactor: move outline-ss-server to cmd/ and hide code (#136)
- Moves the binary to the standard `cmd` directory, and paves the way for having more binaries.
- Hides internal code to prevent external dependencies.
