+++ internal/slicepool/slicepool.go
