+++ internal/slicepool/slicepool_test.go
